fix(provider): maxint for storage and vms (#694)
* fix(provider): use int64 for vm and disk storage


* revert changes for memory


* revert changes for memory


* fix test


* fix int -> int64 conversion for disk size


---------

+++ proxmox/types/disk_size_test.go
@@ -84,7 +84,7 @@ func TestToFromGigabytes(t *testing.T) {
 
 	tests := []struct {
 		name string
-		size int
+		size int64
 		want string
 	}{
 		{"handle 0 size", 0, "0"},
